Several changes to various signal and epoch functions to use vectorized version of `get_epoch_indices` instead of looping through epochs. 

Changes to `average_away_epoch_occurrences` to speed up large loops.

Further speedups in this area will be difficult, but I'm still thinking on it.

Tested with:
```
batch = 308
cell_id = 'AMT003c-19-1'
```
Timing:
```
%%timeit
train, test = rec.split_using_epoch_occurrence_counts('^STIM')
train = average_away_epoch_occurrences(train)
test = average_away_epoch_occurrences(test)
```
Results:
```
Old: 16.5 s ± 1.54 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
New: 4.88 s ± 408 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
~3.4x improvement!
```